why was this such a struggle-finally delete complete and request still working
